% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{austria}
\alias{austria}
\title{Shape file for Austria}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9 rows and 12 columns.
}
\usage{
austria
}
\description{
Shape file for Austria
}
\examples{
\dontrun{
# created from
url <- "https://geodata.ucdavis.edu/gadm/gadm4.0/shp/gadm40_AUT_shp.zip"
austria <- get_shapes(url)
austria <- austria \%>\% thin(tolerance = 0.0001)
usethis::use_data(austria)
}
library(ggplot2)
library(dplyr)
data(austria)
austria \%>\% ggplot() + geom_sf()
}
\keyword{datasets}
