% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymean.r
\name{mymean}
\alias{mymean}
\title{Mymean function}
\usage{
mymean(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric value}

\item{na.rm}{logical value - should missing values be ignored?}
}
\description{
Mymean function
}
