% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_worldle.R
\name{play_worldle}
\alias{play_worldle}
\title{Show the shapefile of a country}
\usage{
play_worldle()
}
\description{
You have up to six guesses to identify the right country
}
\examples{
if (interactive()) {
play_worldle()
}
}
