% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shapes.R
\name{valid_url}
\alias{valid_url}
\title{helper function}
\source{
https://stackoverflow.com/questions/52911812/check-if-url-exists-in-r
}
\usage{
valid_url(url_in, t = 2)
}
\arguments{
\item{url_in}{character value of the website to be checkd for validity}

\item{t}{timeout time in seconds.}
}
\value{
logic value: TRUE for valid connection, FALSE for not
}
\description{
check whether a url is valid
}
\examples{
 valid_url("http://www.stat.iastate.edu")
 valid_url("http://www.not/a/good/url.de")
}
