% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{austria}
\alias{austria}
\title{Shape file for Austria}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
austria
}
\description{
Shape file for Austria
}
\examples{
\dontrun{
# created from
austria <- get_shapes("https://geodata.ucdavis.edu/gadm/gadm3.6/shp/gadm36_AUT_shp.zip")
austria <- thin(austria, 0.001)
use_data(austria, overwrite=TRUE)
}
library(ggplot2)
library(dplyr)
data(austria)
austria \%>\% ggplot() + geom_sf()
}
\keyword{datasets}
