% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymean.R
\name{mymean}
\alias{mymean}
\title{Test Driven Development}
\usage{
mymean(x, na.rm = T)
}
\arguments{
\item{x}{vector of numbers}

\item{na.rm}{boolean, should missing values in vector x be ignored?}
}
\value{
average of x
}
\description{
A small example to see how far we come with implementing test driven development
in a very controlled situation with a definite and simple solution we can check.
}
